requirements.txt
# Core framework
Flask>=3.0.3
Flask-SQLAlchemy>=3.1.1
Flask-Cors>=4.0.0
gunicorn>=21.2.0

# Configuration
python-dotenv>=1.0.1

# Logging
json-log-formatter>=1.0

# Data & scraping
beautifulsoup4>=4.12.3
html5lib>=1.1
lxml>=5.3.0
numpy>=2.1.1
pandas>=2.2.3
yfinance>=0.2.44

# Database
SQLAlchemy>=2.0.32

# Serialization
dataclasses-json>=0.6.0

# HTTP caching & rate limiting
requests-cache>=1.0.0
requests-ratelimiter>=0.5.0
pyrate-limiter>=3.0.0

# Utilities
python-dateutil>=2.9.0.post0
pytz>=2024.1
requests>=2.32.3
